Define a Slate Input

Define a Slate Input

slateInput(
  name,
  type = "character",
  default = NULL,
  parent = "auto",
  long.name = "",
  description = "",
  condition = NULL,
  wizards = list(),
  input.type = type,
  ...
)

Arguments

name
the name of the input.

type
the type of the input. use names(input.handlers) for a list of valid input types.

default
the default value for the input.

parent
the parent of this input. Can be either the name of a group or page, or NULL to have the input be a child of root, or "auto" (the default) to infer the parent based on the order of page, group and input definitions. See inferSlateLayout() for details.

long.name
the name to display in the user interface, or NULL to use the input's name

description
the description of the input. Usually this is shown via a tooltip.

condition
an expression that determines the visibility of this input, or NULL to always show the input.

wizards
a list of wizards associated with this input. Wizards will ususally be displayed in a dropdown menu above the input. See slateWizard() for details.

input.type
legacy: use type instead

...
additional input parameters to be stored in the input structure. These are usually type-specific parameters, such as choices if the input is of type select.

Value

A list of input parameters suitable to be passed along to other rslates functions.