Setup Slate Display and Behaviour Options

Setup Slate Display and Behaviour Options

slateOptions(
  envir = new.env(),
  height = NULL,
  use.card = TRUE,
  card.header = TRUE,
  open.settings = TRUE,
  inputs.style = c("tabset", "collapses", "flowing")
)

Arguments

envir
the parent environment where the slate code is executed.

height
height of the slate in any css unit (e.g. "500px"). If NULL the height of the slate is unrestricted.

use.card
use a bs4 card as a container for the slate UI.

card.header
show the card header. Only used if use.card is TRUE.

open.settings
start with the input settings panel open.

inputs.style
the style used for the inputs panel. "tabset" uses a tabsetPanel as a container for input pages. "collapses" uses bs4 collapse panels. "flowing" displays all pages in a single vertical container.

Value

a list of options to be passed to slateUI() and slateServer().