Create a Modal Dialog

Create a Modal Dialog

slatesModal(
  id,
  session,
  submit.fun,
  ui.fun,
  observers = list(),
  focus.on.show = NULL
)

Arguments

id
id of the modal dialog.

session
server session.

submit.fun
a function without arguments that returns the a list of results of the modal dialog when the user presses OK.

ui.fun
a function taking any number of arguments that returns the ui of the body of the modal.

observers
an optional list of observers to be destroyed when the modal is destroyed.

focus.on.show
unused

Value

a list with the following members. id: the id of the modal, show: a function with signature function(callback, title = NULL, ...) that brings up the modal dialog when called. The callback function is called by the modal dialog when the user presses OK. Each value returned by submit.fun is passed to this function as an argument. Extra ... parameters are passed to the ui.fun function.